msm_serial: Don't configure UART IRQ as wakeup interrupt
Currently UART IRQ is being configured as wakeup interrupt once
UART port is being opened and not unconfigured on closing the port.
It is not required to configure UART IRQ as wakeup interrupt as
wakeup functionality is only supported through UART RX GPIO IRQ line
if it is required and enabled for the use case. Enabling UART IRQ as
wakeup interrupt consumes more power on TCXO state.

CRs-Fixed: 452191
Change-Id: I25f86316fdbd356547e0b6921bde23aff3a56355
Signed-off-by: Saket Saurabh <ssaurabh@codeaurora.org>